Fix ANT init failure
The issue was explored and related to IAR 9.40 vs. 8.10 where the issue does not occur on the exact same repo.  There is significant code size difference in the ant.o files so the compiler is doing things differently.  In troubleshooting, the timing of SEN seemed to be the issue as stepping through functions did not result in the issue.  The fix was to first assign u8ErrorCount to receive the returned value from AntExpectResponse() and this was done when u8ErrorCount was still bErrorStatus.  That worked, but AntExpectResponse returns u8 so there was a compiler warning on type mismatch.  Thus u8ErrorCount replaced bErrorStatus.  The leds.c change was from the "typical" mysterious phantom hard faults that sometimes occur.  Neither fix definitively identifies root cause.